Zylbercweig/zibn-shtern/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zibn-shtern"
version = "0.1.0"
description = "Audit and normalization pipeline for place names from Zylbercweig's Lexicon of Yiddish Theater"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "pyyaml>=6.0",
  "rapidfuzz>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
